Rename files to *EventLinks
